Correction des warnings

+++ src/string_manip.cpp
@@ -8,7 +8,7 @@ bool ends_with(std::string const & value, std::string const & ending)
 }
 
 void replace(std::string* str, std::string old_part, std::string new_part) {
-    for(int i = 0; i < str->size(); i++) {
+    for(long unsigned int i = 0; i < str->size(); i++) {
         if(std::equal(old_part.begin(), old_part.end(), str->begin() + i)) {
             str->replace(i, old_part.size(), new_part);
         };
